refactor(handler): split handler.go, fix concurrency, add new handlers
Split handler.go (1615 lines) into focused files within the same
package: types.go, pumps.go, heartbeat.go, certs.go, notifications.go,
and dispatch.go. The resulting handler.go (622 lines) contains only
the Handler struct, New constructor, connection lifecycle (Connect,
Run, Close), and the Send/SendRaw API. No public API or logic
changes from the split itself.

Concurrency fixes applied along the way:
- Add buffered semaphore (cap 64) in readPump to bound concurrent
  message handlers and prevent goroutine exhaustion.
- Wrap every dispatcher call in context.WithTimeout (30s default,
  10 min for known long-running actions) with panic recovery; the
  worker goroutine uses non-blocking sends on a buffered outcome
  channel so it always exits, even after the parent returns on
  timeout.
- Replace context.Background() in ScheduleReboot with the handler
  shutdown channel so the reboot is cancelled on Close.
- Switch renewCertificates to SharedTransport().Clone() with mTLS
  override to stop creating a fresh HTTP client every 24 h.
- Log conn.Close() errors instead of silently dropping them.

Add a JSON nesting-depth validator (MaxJSONDepth = 32) that runs on
every inbound WebSocket message before json.Unmarshal, dropping
messages that exceed the limit.

Migrate HeartbeatStats, HeartbeatDisk, and HeartbeatNetworkIO counter
fields from uint64 to int64 to avoid overflow when parsed as Python
int on the backend.

Add new WebSocket handlers:
- list_processes: enumerates running processes via gopsutil and
  returns PID, PPID, name, CPU%, memory, status, username, create
  time, and truncated cmdline.
- set_maintenance_mode / get_maintenance_status: maintenance state
  tracked via atomic.Bool, exposed through IsInMaintenance().

+++ internal/handler/heartbeat_helpers.go
@@ -17,9 +17,9 @@ func buildDiskStats(stats *monitor.Stats) []HeartbeatDisk {
 		diskStats = append(diskStats, HeartbeatDisk{
 			Device:      d.Device,
 			Mountpoint:  d.Mountpoint,
-			Total:       d.Total,
-			Used:        d.Used,
-			Free:        d.Free,
+			Total:       int64(d.Total),
+			Used:        int64(d.Used),
+			Free:        int64(d.Free),
 			UsedPercent: d.UsedPercent,
 		})
 	}
@@ -36,10 +36,10 @@ func aggregateNetworkIO(stats *monitor.Stats) *HeartbeatNetworkIO {
 		totalPacketsRecv += n.PacketsRecv
 	}
 	return &HeartbeatNetworkIO{
-		BytesSent:   totalBytesSent,
-		BytesRecv:   totalBytesRecv,
-		PacketsSent: totalPacketsSent,
-		PacketsRecv: totalPacketsRecv,
+		BytesSent:   int64(totalBytesSent),
+		BytesRecv:   int64(totalBytesRecv),
+		PacketsSent: int64(totalPacketsSent),
+		PacketsRecv: int64(totalPacketsRecv),
 	}
 }
 
@@ -52,11 +52,11 @@ func (h *Handler) buildBaseHeartbeat(ctx context.Context, stats *monitor.Stats) 
 		Stats: HeartbeatStats{
 			CPUPercent:    stats.CPU.UsagePercent,
 			MemoryPercent: stats.Memory.UsedPercent,
-			MemoryUsed:    stats.Memory.Used,
-			MemoryTotal:   stats.Memory.Total,
+			MemoryUsed:    int64(stats.Memory.Used),
+			MemoryTotal:   int64(stats.Memory.Total),
 			Disk:          buildDiskStats(stats),
 			NetworkIO:     aggregateNetworkIO(stats),
-			UptimeSeconds: stats.Uptime,
+			UptimeSeconds: int64(stats.Uptime),
 			ProcessCount:  stats.ProcessCount,
 			Timezone:      stats.Timezone,
 		},
